+++ src-tauri/src/commands/mastery_tracker/get_items.rs
@@ -1,8 +1,8 @@
-use tauri::State;
 use crate::database::db::UserDb;
 use crate::database::repositories::mastery_repo;
 use crate::models::database::filters::MasteryFilters;
 use crate::models::database::item::Item;
+use tauri::State;
 
 #[tauri::command]
 pub async fn get_items(