+++ src-tauri/src/commands/task_tracker/get_tasks.rs
@@ -1,8 +1,8 @@
-use tauri::State;
 use crate::database::db::UserDb;
 use crate::database::repositories::task_repo;
 use crate::models::database::filters::TaskFilters;
 use crate::models::database::task::Task;
+use tauri::State;
 
 #[tauri::command]
 pub async fn get_tasks(