+++ src-tauri/src/commands/task_tracker/set_task.rs
@@ -1,9 +1,9 @@
-use tauri::State;
 use crate::database::db::UserDb;
 use crate::database::repositories::task_repo;
 use crate::database::services::task_services::{ get_period_start, ResetType };
 use crate::models::database::task::Task;
 use chrono::Utc;
+use tauri::State;
 
 #[tauri::command]
 pub async fn set_task(state: State<'_, UserDb>, id: String, count: i32) -> Result<Task, String> {